LayoutTests/http/tests/cache/disk-cache/speculative-validation/cacheable-redirect-expected.txt

Tests that speculative validation does not break redirects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS frames[0].getComputedStyle(testDiv).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
